Fixed issue with Draw_BeginDisc being called when there was not, in fact, a disc image to be drawn on screen. Removed duplicate audio buffer creation in both MacOS and the experimental iOS versions. Fixed issue with miscalculated texture coordinates for surfaces containing the data for skyboxes.

+++ id1/r_sky.cpp
@@ -781,8 +781,8 @@ void R_SetSkyBox (float rotate, vec3_t axis)
 		std::string prefix = suf[r_skysideimage[i]];
 		std::string path = "gfx/env/" + r_skyboxprefix + prefix + ".tga";
 		R_LoadSkyImage(path, prefix, r_skytexinfo[i].texture);
-        r_skyfaces[i].texturemins[0] = -r_skytexinfo[i].texture->width / 2;
-        r_skyfaces[i].texturemins[1] = -r_skytexinfo[i].texture->height / 2;
+        r_skyfaces[i].texturemins[0] = -(int)r_skytexinfo[i].texture->width / 2;
+        r_skyfaces[i].texturemins[1] = -(int)r_skytexinfo[i].texture->height / 2;
         r_skyfaces[i].extents[0] = r_skytexinfo[i].texture->width;
         r_skyfaces[i].extents[1] = r_skytexinfo[i].texture->height;
 	}
